bets_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
asyncpg
python-jose[cryptography]
httpx
